refactor api package to server
