+++ internal/server/server.go
@@ -1,6 +1,6 @@
-// Package api handles all logic around the HTTP server and the
+// Package server handles all logic around the HTTP server and the
 // coordination of all database connections.
-package api
+package server
 
 import (
 	"log"
@@ -16,17 +16,17 @@ import (
 const requestTimeout time.Duration = 10 * time.Second
 const maxHeaderSize = 4096
 
-// ShortenerAPI is the wrapper for the main HTTP server.
-type ShortenerAPI struct {
+// Server is the wrapper for the main HTTP server.
+type Server struct {
 	server   *http.Server
 	Database database.Connection
 	Cache    cache.Connection
 	Auth     auth.Connection
 }
 
-// NewShortenerAPI constructs a new shortener API instance.
-func NewShortenerAPI() ShortenerAPI {
-	api := ShortenerAPI{
+// NewServer constructs a new shortener API instance.
+func NewServer() Server {
+	api := Server{
 		server:   nil,
 		Database: nil,
 		Cache:    nil,
@@ -37,12 +37,12 @@ func NewShortenerAPI() ShortenerAPI {
 }
 
 // CreateURLRoute creates a new shortened URL.
-func (api *ShortenerAPI) CreateURLRoute(w http.ResponseWriter, _ *http.Request) {
+func (api *Server) CreateURLRoute(w http.ResponseWriter, _ *http.Request) {
 	w.WriteHeader(http.StatusOK)
 }
 
 // GetURLRoute fetches a full URL using its shortened ID.
-func (api *ShortenerAPI) GetURLRoute(w http.ResponseWriter, r *http.Request) {
+func (api *Server) GetURLRoute(w http.ResponseWriter, r *http.Request) {
 	id := strings.ToUpper(r.URL.Path[1:])
 
 	_, err := w.Write([]byte(id))
@@ -52,7 +52,7 @@ func (api *ShortenerAPI) GetURLRoute(w http.ResponseWriter, r *http.Request) {
 }
 
 // SetupRoutes constructs a serve mux and mounts all routes to it.
-func (api *ShortenerAPI) SetupRoutes() *http.ServeMux {
+func (api *Server) SetupRoutes() *http.ServeMux {
 	mux := http.NewServeMux()
 	mux.HandleFunc("POST /", api.CreateURLRoute)
 	mux.HandleFunc("GET /", api.GetURLRoute)
@@ -61,14 +61,14 @@ func (api *ShortenerAPI) SetupRoutes() *http.ServeMux {
 }
 
 // Listen launches the HTTP server under the given address.
-func (api *ShortenerAPI) Listen(addr string) bool {
+func (api *Server) Listen(addr string) bool {
 	api.configureServer(addr)
 	err := api.server.ListenAndServe()
 
 	return err == nil
 }
 
-func (api *ShortenerAPI) configureServer(addr string) {
+func (api *Server) configureServer(addr string) {
 	api.server = &http.Server{
 		ReadTimeout:                  requestTimeout,
 		ReadHeaderTimeout:            requestTimeout,
